Rename createResdientPermit components folder to residentPermit
The components are shared by both creating and editing resident permit

Refs: PV-229
