+++ src/components/residentPermit/PermitInfo.module.scss
