+++ src/components/residentPermit/VehicleInfo.module.scss
